feat: ajustado o model do ProdutoInvestimento e populado o banco além de criar a controller que retorna os produtos cadastrador.

+++ src/SimulaCred.Domain/Entities/ProdutoInvestimento.cs
@@ -6,11 +6,11 @@ public class ProdutoInvestimento: BaseEntity
 {
     public string Nome { get; init; }
     public ProdutoInvestimentoTipo Tipo { get; init; }
-    public float Rentabilidade { get; init; }
+    public decimal Rentabilidade { get; init; }
     public ProdutoInvestimentoRisco Risco { get; init; }
     bool Ativo { get; init; }
 
-    public ProdutoInvestimento(string nome, ProdutoInvestimentoTipo tipo, float rentabilidade, ProdutoInvestimentoRisco risco)
+    public ProdutoInvestimento(string nome, ProdutoInvestimentoTipo tipo, decimal rentabilidade, ProdutoInvestimentoRisco risco)
     {
         Nome = nome;
         Tipo = tipo;
